fix(flights): eliminate two -race failures + green CI gates for #475
Two data races (ryanair pkg-global base URL vs live concurrent search;
afklm shared-slice backing array across goroutines) fixed with an
atomic.Value guard and per-goroutine request builders. errcheck fix on
a bare os.Unsetenv.

Instead of allowlisting the two files that grew past the 800-line
ceiling, split them into cohesive same-package files:
  internal/flights/roundtrip.go   861 -> 375 (+ roundtrip_native.go)
  internal/hacks/rail_fly_test.go 813 -> 558 (+ rail_fly_cost_test.go)

Instead of the skip-regression-test label, added real same-package
regression tests for the three fix-touched packages:
  internal/models  hacksaving candidate scoring
  internal/profile SK alliance (SkyTeam) resolution
  mcp              SK frequent-flyer card -> Flying Blue, not EuroBonus

Gates verified locally (GOTOOLCHAIN=go1.26.5): gofmt clean, go build,
go vet, go test -race across flights/afklm/models/profile/hacks/mcp all
pass; scripts/ci/check-file-size.sh exit 0 with only CHANGELOG allowlisted.
